Refactor: Retirar my_expenses_page do main e criar page própria
